+++ tests/test_runtime_errors.py
@@ -159,11 +159,12 @@ def test_unable_to_parse(capsys: "CaptureFixture") -> None:
     """Check error message shows if we're unable to parse notebook."""
     path = Path("tests") / "data/invalid_notebook.ipynb"
     path.write_text("foo")
-    main(["flake8", str(path)])
+    result = main(["flake8", str(path)])
     path.unlink()
     message = "nbQA failed to process"
     _, err = capsys.readouterr()
     assert message in err
+    assert result == 123
 
 
 def test_unable_to_parse_with_valid_notebook(capsys: "CaptureFixture") -> None: