+++ nbqa/__init__.py
@@ -1,3 +1,3 @@
 """Attributes used by docs / packaging."""
 
-__version__ = "1.5.0"
+__version__ = "1.5.1"