+++ docs/history.rst
@@ -5,6 +5,12 @@ Changelog
 1.5.0 (???)
 ~~~~~~~~~~~
 ``nbqa`` now removes empty cells which were not empty to begin with.
+Markdown files saved via ``Jupytext`` will now be processed as well
+(thanks @basnijholt and @rgommers for the suggestion!)
+If ``nbqa`` is passed an invalid notebook, it will exit 123. If it's
+passed a non-Python notebook, it'll exit 0.
+Don't try to process files with the wrong extensions, even if passed
+explicitly.
 Added support for subcommands (thanks @dnoliver for the issue, @s-weigand for the fix)
 
 1.4.0 (2022-07-17)